bug_report: Request information about previously working version.
Currently we receive many reports that do not include any information at
all about previously working versions. While we are generally able to
connect a new issue to a change merged recently it is much easier for us
to receive this information in the report directly, instead of having to
ask for it.
Besides new contributors might be incentivized to perform research on
their own when asked for such information, perhaps resolving the issue
when the change is small.